requirements.txt
azure-cognitiveservices-speech
openai
Flask